requirements.txt
google-cloud-pubsub
avro
requests
google-api-core